Simplify ProposedMultisigBundle
The current ProposedMultisigBundle validates various conditions and if
applicable, raises exceptions. At a brief glance, the code did not look
as human friendly as it can be.

The proposed changes seek to simply the module as well as breakup the
code into logical pieces which makes it easy to reason with. As a
consequence, the code is also much more compact.